Move ScriptEditorExamples to scripeditor/examples.py
